Large Refactor, added modified path and loop enumeration with carry over bit

+++ ValidPathRowEnumerator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics.Contracts;
+using System.Numerics;
 using System.Threading.Tasks;
 
 namespace CrawfisSoftware.PCG
@@ -287,7 +288,7 @@ namespace CrawfisSoftware.PCG
         /// <param name="width">The width of the row</param>
         /// <param name="row">The inflow bit pattern</param>
         /// <returns></returns>
-        public static List<int> InflowsFromBits(int width, int row)
+        public static List<int> InflowsFromBits(int width, BigInteger row)
         {
             var inFlows = new List<int>();
             int mask = 1;
